refactor: model field
